style: 테스트 타겟에 있던 Qeueu 타입 모델로 이동  #2
